requirements.txt
# Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.11

# Machine learning
scikit-learn>=1.3
yellowbrick>=1.5

# Deep learning
tensorflow>=2.13

# NLP / text
# (TF-IDF and SVM are covered by scikit-learn)

# Time series / financial data
yfinance>=0.2

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15

# Image processing
Pillow>=10.0

# Jupyter
notebook>=7.0
ipykernel>=6.0
